avutil: Cast AV_NOPTS_VALUE to int64_t explicitly
Thus comparsion against int64_t value will not raise warning
(from -Wextra set) about comparsion of unsigned and signed integer

commiter added () and changed the litteral to unsigned

+++ libavutil/avutil.h
@@ -223,7 +223,7 @@ const char *av_get_media_type_string(enum AVMediaType media_type);
  * either pts or dts.
  */
 
-#define AV_NOPTS_VALUE          INT64_C(0x8000000000000000)
+#define AV_NOPTS_VALUE          ((int64_t)UINT64_C(0x8000000000000000))
 
 /**
  * Internal time base represented as integer
